Update AbstractPaynlResponse.php
When cancelling a transaction at the pay.nl website you'll get a succesful transaction status and thus the transaction is marked as paid. I've added a check to make sure that the paymentDetails.state is 100 (succesful transaction). If it is not 100 then OmniPay will pick up the transaction status and handle the rest of the process.